requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic-settings>=2.0.3
pydantic-extra-types>=2.1.0
email-validator>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
python-dotenv>=1.0.0
alembic>=1.12.1
psycopg2-binary>=2.9.9
